requirements.txt
beautifulsoup4==4.3.2
Django==1.6.10
South==1.0.2
coverage==3.7.1
django-appconf==1.0.1
django-autocomplete-light==2.0.9
django-bootstrap-pagination==1.5.1
django-braces==1.4.0
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-extensions==1.5.0
django-geoposition==0.2.2
django-model-utils==2.2
django-twitter-bootstrap==3.3.0
djet==0.0.10
frogress==0.9.1
gunicorn==19.2.1
ipython==2.4.1
lxml==3.4.2
mock==1.0.1
psycopg2==2.6
pyaavso==0.1.4
pyephem==3.7.5.3
pygcvs==0.1.4
raven==5.2.0
requests==2.5.1
six==1.9.0
